+++ python_linq/__init__.py
@@ -1,2 +1,5 @@
-from python_linq.core import From, Grouping, Joining, NoSuchElementError
-from python_linq.linq_exceptions import *
+from ._core import From, Grouping, Joining
+from ._linq_exceptions import NoSuchElementError
+
+
+__all__ = ["From", "Grouping", "Joining", "NoSuchElementError"]
