+++ python_linq/_linq_exceptions.py
